fixing react bug
should give correct location
